Started upgrade to C# 6 and VS 2015 - Fixed some style issues - Added CFS (work in progress)

+++ Applications/D3Theory/Data/SimulationSampleSet.cs
@@ -153,7 +153,7 @@
             }
 
             float total = this.DamageTotal.ContainsKey(damageType) ? this.DamageTotal[damageType] : 0f;
-            return total / (float)seconds;
+            return total / seconds;
         }
 
         // -------------------------------------------------------------------
